+++ CHANGELOG.md
@@ -10,6 +10,7 @@ History of the project TipPop.
 | [Component] | New component added or old component refactored         |
 | [Bugfix]    | Fix for any github issue. Mention issue link            |
 | [Docfix]    | Changes in documentation. Mention document section link |
+| [Style]     | Changes in styling. Mention component                   |
 
 ## Changes
 
@@ -17,3 +18,14 @@ History of the project TipPop.
 
 - [Setup] - Initial project baseline setup using CRA.
 - [Component] - Add TipPop base component and example in main App component.
+
+### [0.0.2]
+
+- [Component] - Convert TipPop to HOC for tooltip trigger to be passed as child.
+- [Component] - Add toggling behavior logic to TipPop for click and hover.
+- [Style] - Add basic style for tooltips in TipPop container.
+- [Docfix] - Add jsdoc and proptypes to TipPop for ease of usability
+- [Docfix] - Add examples and styling to App component.
+- [Docfix] - Add examples for demo for behavior and styiling customizations
+- [Docfix] - Add documentation for TipPop component package in readme.
+- [Style] - Add tooltip trigger container custom styling.